moving requests to dedicated folder + moving storage to dedicated folder + DRY code
